# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/api/v1/admin/card": {
      "get": {
        "tags": [
          "AdminControllerApi"
        ],
        "summary": "Получить список карт",
        "description": "Этот метод позволяет получить список карт общий, или по конкретному пользователю\nс фильтрацией по статусам",
        "operationId": "getCards",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10,
              "minimum": 1
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "ACTIVE",
                  "BLOCKED",
                  "EXPIRED"
                ]
              }
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageCardSmallResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные запроса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageCardSmallResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "Карты успешно получены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageCardSmallResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не существует",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageCardSmallResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "AdminControllerApi"
        ],
        "summary": "Создание карты пользователю",
        "description": "Этот метод позволяет создать карту для пользователя",
        "operationId": "createCard",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные запроса",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "200": {
            "description": "Карта успешно создана",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/admin/block/{id}": {
      "post": {
        "tags": [
          "AdminControllerApi"
        ],
        "summary": "Заблокировать карту",
        "description": "Этот метод позволяет заблокировать карту пользователя",
        "operationId": "blockCard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Карта не найдена"
          },
          "200": {
            "description": "Карта успешно заблокирована"
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          },
          "400": {
            "description": "Некорректные данные запроса"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/admin/active/{id}": {
      "post": {
        "tags": [
          "AdminControllerApi"
        ],
        "summary": "Активировать карту",
        "description": "Этот метод позволяет активировать карту пользователя, которая заблокирована или просрочена",
        "operationId": "activeCard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "expiryDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Карта успешно активирована"
          },
          "404": {
            "description": "Карта не найдена"
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          },
          "400": {
            "description": "Некорректные данные запроса"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/card": {
      "get": {
        "tags": [
          "CardControllerApi"
        ],
        "summary": "Получить список моих карт",
        "description": "Этот метод позволяет получить мои карты",
        "operationId": "getMyCards",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10,
              "minimum": 1
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "ACTIVE",
                  "BLOCKED",
                  "EXPIRED"
                ]
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageCardSmallResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные запроса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageCardSmallResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "Карты успешно получены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageCardSmallResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/admin/requestsToBlock": {
      "get": {
        "tags": [
          "AdminControllerApi"
        ],
        "summary": "Получить список запросов на блокировку",
        "description": "Этот метод позволяет получить список запросов на блокировку карт",
        "operationId": "getRequests",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Запросы успешно получены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageRequestToBlockResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageRequestToBlockResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные запроса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageRequestToBlockResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CardRequestDto": {
        "type": "object",
        "properties": {
          "cardNumber": {
            "type": "string"
          },
          "expiryDate": {
            "type": "string",
            "format": "date"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "Pageablenull": {
        "type": "object",
        "properties": {
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/Sortnull"
          },
          "paged": {
            "type": "boolean"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "Sortnull": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
...
      "CardSmallResponseDto": {
        "type": "object",
        "properties": {
          "cardId": {
            "type": "string",
            "format": "uuid"
          },
          "cardNumber": {
            "type": "string"
          },
          "cardStatus": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "BLOCKED",
              "EXPIRED"
            ]
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "PageCardSmallResponseDto": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/Pageablenull"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardSmallResponseDto"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/Sortnull"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "PageRequestToBlockResponseDto": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/Pageablenull"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RequestToBlockResponseDto"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/Sortnull"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "RequestToBlockResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cardId": {
            "type": "string",
            "format": "uuid"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}